requirements.txt
Flask==2.0.2
pycryptodome==3.10.1
requests==2.26.0
secrets
